requirements (1).txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0
Pillow>=8.0.0
tqdm>=4.50.0
lpips>=0.1.4
